+++ src/main/java/MSGUserService/models/entities/UserEntity.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -18,7 +19,7 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;
+    private BigDecimal id;
 
     @Column(name = "username")
     private String username;
@@ -27,7 +28,7 @@ public class UserEntity {
     private String passwordDigest;
 
     @Column(name = "user_code")
-    private long userCode;
+    private BigDecimal userCode;
 
     @Column(name = "email")
     private String email;
